+++ openghg/retrieve/_access.py
@@ -1,9 +1,9 @@
 import json
 import logging
 from io import BytesIO
-from typing import Any, Dict, List, Optional, Union
-from openghg.types import SearchError
+from typing import Any, Optional, Union
 
+from openghg.dataobjects._basedata import _BaseData  # TODO: expose this type?
 from openghg.dataobjects import (
     BoundaryConditionsData,
     FluxData,
@@ -11,7 +11,8 @@ from openghg.dataobjects import (
     ObsColumnData,
     ObsData,
 )
-from openghg.util import decompress, decompress_str, hash_bytes, running_on_hub
+from openghg.types import SearchError
+from openghg.util import combine_and_elevate_inlet, decompress, decompress_str, hash_bytes, running_on_hub
 from pandas import Timestamp
 from xarray import Dataset, load_dataset
 
@@ -20,21 +21,19 @@ logger.setLevel(logging.DEBUG)  # Have to set level for logger as well as handle
 
 DataTypes = Union[BoundaryConditionsData, FluxData, FootprintData, ObsColumnData, ObsData]
 multDataTypes = Union[
-    List[BoundaryConditionsData], List[FluxData], List[FootprintData], List[ObsColumnData], List[ObsData]
+    list[BoundaryConditionsData], list[FluxData], list[FootprintData], list[ObsColumnData], list[ObsData]
 ]
 
 
 def _get_generic(
-    sort: bool = True,
-    elevate_inlets: bool = False,
+    combine_multiple_inlets: bool = False,
     ambig_check_params: Optional[list] = None,
     **kwargs: Any,
-) -> Any:
+) -> _BaseData:
     """Perform a search and create a dataclass object with the results if any are found.
 
     Args:
         data_class: Type of dataobject to create
-        sort: Sort Dataset during recombination
         elevate_inlets: Elevate the inlet attribute to be a variable within the Dataset
         ambig_check_params: Parameters to check and print if result is ambiguous.
         kwargs: Additional search terms
@@ -42,7 +41,6 @@ def _get_generic(
         dataclass
     """
     from openghg.retrieve import search
-    from openghg.store.spec import define_data_type_classes
 
     results = search(**kwargs)
 
@@ -52,49 +50,40 @@ def _get_generic(
         logger.exception(err_msg)
         raise SearchError(err_msg)
 
-    if sort:
-        logger.warning(
-            "Sorting results by time has been moved to the data method of ObsData, please use that instead."
-        )
-
-    if elevate_inlets:
-        logger.warning(
-            "Elevating inlet functionality has been moved to the data method of ObsData, please use that instead."
-        )
-
     # TODO: UPDATE THIS - just use retrieve when retrieve_all is removed.
-    retrieved_data: Any = results.retrieve_all()
+    retrieved_data = results.retrieve_all()
 
     if retrieved_data is None:
         err_msg = f"Unable to retrieve results for {keyword_string}"
         logger.exception(err_msg)
         raise SearchError(err_msg)
     elif isinstance(retrieved_data, list) and len(retrieved_data) > 1:
-        param_diff_formatted = _metadata_difference_formatted(data=retrieved_data, params=ambig_check_params)
-        err_msg = f"""
-        Multiple entries found for input parameters for {keyword_string}.
-        Parameter differences:
-        {param_diff_formatted}
-        Please supply additional parameters or set ranking.
-        """
-        logger.exception(err_msg)
-        raise SearchError(err_msg)
+        if combine_multiple_inlets:
+            result = combine_and_elevate_inlet(retrieved_data)
+        else:
+            param_diff_formatted = _metadata_difference_formatted(
+                data=retrieved_data, params=ambig_check_params
+            )
+            err_msg = f"""
+            Multiple entries found for input parameters for {keyword_string}.
+            Parameter differences:
+            {param_diff_formatted}
+            Please supply additional parameters or set ranking.
+            """
+            logger.exception(err_msg)
+            raise SearchError(err_msg)
     elif isinstance(retrieved_data, list):
-        retrieved_data = retrieved_data[0]
+        result = retrieved_data[0]
+    else:
+        result = retrieved_data
 
-    # TODO: Included output of this as Any for now because we there are many
-    # options for types returned but can update this
-
-    # We can only get a single data object back here but mypy doesn't understand that
-    # retrieved_data = cast(ObsData, retrieved_data)
-
-    return retrieved_data
+    return result
 
 
 def get_obs_surface(
     site: str,
     species: str,
-    inlet: Optional[str] = None,
+    inlet: Optional[Union[str, slice]] = None,
     height: Optional[str] = None,
     start_date: Optional[Union[str, Timestamp]] = None,
     end_date: Optional[Union[str, Timestamp]] = None,
@@ -115,7 +104,9 @@ def get_obs_surface(
         species: Species identifier e.g. ch4 for methane.
         start_date: Output start date in a format that Pandas can interpret
         end_date: Output end date in a format that Pandas can interpret
-        inlet: Inlet height above ground level in metres
+        inlet: Inlet height above ground level in metres; This can be a single value or `slice(lower, upper)`
+            can be used to search for a range of values. `lower` and `upper` can be int, float, or strings
+            such as '100m'.
         height: Alias for inlet
         average: Averaging period for each dataset. Each value should be a string of
         the form e.g. "2H", "30min" (should match pandas offset aliases format).
@@ -138,7 +129,7 @@ def get_obs_surface(
 
     if running_on_hub():
         raise NotImplementedError("Cloud functionality marked for rewrite.")
-        to_post: Dict[str, Union[str, Dict]] = {}
+        to_post: dict[str, Union[str, dict]] = {}
 
         to_post["function"] = "get_obs_surface"
 
@@ -208,7 +199,7 @@ def get_obs_surface(
 def get_obs_surface_local(
     site: str,
     species: str,
-    inlet: Optional[str] = None,
+    inlet: Optional[Union[str, slice]] = None,
     height: Optional[str] = None,
     start_date: Optional[Union[str, Timestamp]] = None,
     end_date: Optional[Union[str, Timestamp]] = None,
@@ -231,7 +222,9 @@ def get_obs_surface_local(
         species: Species identifier e.g. ch4 for methane.
         start_date: Output start date in a format that Pandas can interpret
         end_date: Output end date in a format that Pandas can interpret
-        inlet: Inlet height above ground level in metres
+        inlet: Inlet height above ground level in metres; This can be a single value or `slice(lower, upper)`
+            can be used to search for a range of values. `lower` and `upper` can be int, float, or strings
+            such as '100m'.
         height: Alias for inlet
         average: Averaging period for each dataset. Each value should be a string of
         the form e.g. "2H", "30min" (should match pandas offset aliases format).
@@ -244,10 +237,13 @@ def get_obs_surface_local(
         ObsData or None: ObsData object if data found, else None
     """
     import numpy as np
-    from openghg.retrieve import search_surface
-    from openghg.util import clean_string, format_inlet, load_json, synonyms, timestamp_tzaware, get_site_info
+    from openghg.util import (
+        format_inlet,
+        get_site_info,
+        synonyms,
+        timestamp_tzaware,
+    )
     from pandas import Timedelta
-    from openghg.util import synonyms
 
     if running_on_hub():
         raise ValueError(
@@ -287,8 +283,7 @@ def get_obs_surface_local(
     # # Get the observation data
     # obs_results = search_surface(**surface_keywords)
     retrieved_data = _get_generic(
-        sort=True,
-        elevate_inlets=False,
+        combine_multiple_inlets=isinstance(inlet, slice),  # if range passed for inlet, try to combine
         ambig_check_params=["inlet", "network", "instrument"],
         **surface_keywords,  # type: ignore
     )
@@ -373,7 +368,7 @@ def get_obs_surface_local(
         ds_resampled.attrs["averaged_period_str"] = average
 
         # For some variables, need a different type of resampling
-        data_variables: List[str] = [str(v) for v in data.variables]
+        data_variables: list[str] = [str(v) for v in data.variables]
 
         for var in data_variables:
             if "repeatability" in var:
@@ -415,7 +410,7 @@ def get_obs_surface_local(
         data = ds_resampled
 
     # Rename variables
-    rename: Dict[str, str] = {}
+    rename: dict[str, str] = {}
 
     data_variables = [str(v) for v in data.variables]
     for var in data_variables:
@@ -468,6 +463,7 @@ def get_obs_surface_local(
 
 def get_obs_column(
     species: str,
+    max_level: int,
     satellite: Optional[str] = None,
     domain: Optional[str] = None,
     selection: Optional[str] = None,
@@ -479,8 +475,7 @@ def get_obs_column(
     end_date: Optional[Union[str, Timestamp]] = None,
     **kwargs: Any,
 ) -> ObsColumnData:
-    """
-    Extract available column data from the object store using keywords.
+    """Extract available column data from the object store using keywords.
 
     Args:
         species: Species name
@@ -494,7 +489,6 @@ def get_obs_column(
         ObsColumnData: ObsColumnData object
     """
     obs_data = _get_generic(
-        sort=False,
         species=species,
         satellite=satellite,
         domain=domain,
@@ -507,6 +501,71 @@ def get_obs_column(
         end_date=end_date,
         data_type="column",
         **kwargs,
+    )
+
+    if max_level > max(obs_data.data.lev.values) + 1:
+        logger.warning(
+            f"passed max level is above max level in data ({max(obs_data.data.lev.values)+1}). Defaulting to highest level"
+        )
+        max_level = max(obs_data.data.lev.values) + 1
+
+    ## processing taken from acrg/acrg/obs/read.py get_gosat()
+    lower_levels = list(range(0, max_level))
+
+    prior_factor = (
+        obs_data.data.pressure_weights[dict(lev=list(lower_levels))]
+        * (1.0 - obs_data.data.xch4_averaging_kernel[dict(lev=list(lower_levels))])
+        * obs_data.data.ch4_profile_apriori[dict(lev=list(lower_levels))]
+    ).sum(dim="lev")
+
+    upper_levels = list(range(max_level, len(obs_data.data.lev.values)))
+    prior_upper_level_factor = (
+        obs_data.data.pressure_weights[dict(lev=list(upper_levels))]
+        * obs_data.data.ch4_profile_apriori[dict(lev=list(upper_levels))]
+    ).sum(dim="lev")
+
+    obs_data.data["mf_prior_factor"] = prior_factor
+    obs_data.data["mf_prior_upper_level_factor"] = prior_upper_level_factor
+    obs_data.data["mf"] = (
+        obs_data.data.xch4 - obs_data.data.mf_prior_factor - obs_data.data.mf_prior_upper_level_factor
+    )
+    obs_data.data["mf_repeatability"] = obs_data.data.xch4_uncertainty
+
+    # rt17603: 06/04/2018 Added drop variables to ensure lev and id dimensions are also dropped, Causing problems in footprints_data_merge() function
+    drop_data_vars = [
+        "xch4",
+        "xch4_uncertainty",
+        "lon",
+        "lat",
+        "ch4_profile_apriori",
+        "xch4_averaging_kernel",
+        "pressure_levels",
+        "pressure_weights",
+        "exposure_id",
+    ]
+    drop_coords = ["lev", "id"]
+
+    for dv in drop_data_vars:
+        if dv in obs_data.data.data_vars:
+            obs_data.data = obs_data.data.drop_vars(dv)
+    for coord in drop_coords:
+        if coord in obs_data.data.coords:
+            obs_data.data = obs_data.data.drop_vars(coord)
+
+    obs_data.data = obs_data.data.sortby("time")
+
+    obs_data.data.attrs["max_level"] = max_level
+    if species.upper() == "CH4":
+        # obs_data.data.mf.attrs["units"] = "1e-9"
+        obs_data.data.attrs["species"] = "CH4"
+    if species.upper() == "CO2":
+        # obs_data.data.mf.attrs["units"] = "1e-6"
+        obs_data.data.attrs["species"] = "CO2"
+
+    # obs_data.data.attrs["scale"] = "GOSAT"
+
+    obs_data.metadata["transforms"] = (
+        f"For creating mole fraction, used apriori data for levels above max_level={max_level}"
     )
 
     return ObsColumnData(data=obs_data.data, metadata=obs_data.metadata)
@@ -524,8 +583,7 @@ def get_flux(
     time_resolution: Optional[str] = None,
     **kwargs: Any,
 ) -> FluxData:
-    """
-    The flux function reads in all flux files for the domain and species as an xarray Dataset.
+    """The flux function reads in all flux files for the domain and species as an xarray Dataset.
     Note that at present ALL flux data is read in per species per domain or by emissions name.
     To be consistent with the footprints, fluxes should be in mol/m2/s.
 
@@ -541,7 +599,6 @@ def get_flux(
         FluxData: FluxData object
     """
     em_data = _get_generic(
-        sort=False,
         species=species,
         source=source,
         domain=domain,
@@ -574,8 +631,7 @@ def get_bc(
     end_date: Optional[Union[str, Timestamp]] = None,
     **kwargs: Any,
 ) -> BoundaryConditionsData:
-    """
-    Get boundary conditions for a given species, domain and bc_input name.
+    """Get boundary conditions for a given species, domain and bc_input name.
 
     Args:
         species: Species name
@@ -589,7 +645,6 @@ def get_bc(
         BoundaryConditionsData: BoundaryConditionsData object
     """
     bc_data = _get_generic(
-        sort=False,
         species=species,
         bc_input=bc_input,
         domain=domain,
@@ -613,8 +668,7 @@ def get_footprint(
     species: Optional[str] = None,
     **kwargs: Any,
 ) -> FootprintData:
-    """
-    Get footprints from one site.
+    """Get footprints from one site.
 
     Args:
         site: The name of the site given in the footprints. This often matches
@@ -636,7 +690,7 @@ def get_footprint(
     Returns:
         FootprintData: FootprintData dataclass
     """
-    from openghg.util import clean_string, synonyms, format_inlet
+    from openghg.util import clean_string, format_inlet, synonyms
 
     # Find the correct synonym for the passed species
     if species is not None:
@@ -704,10 +758,7 @@ def _scale_convert(data: Dataset, species: str, to_scale: str) -> Dataset:
 
     converter = scale_converter.loc[row]
 
-    if to_scale == converter["scale1"]:
-        direction = "2to1"
-    else:
-        direction = "1to2"
+    direction = "2to1" if to_scale == converter["scale1"] else "1to2"
 
     # flake8: noqa: F841
     # scale_convert file has variable X in equations, so let's create it
@@ -721,8 +772,7 @@ def _scale_convert(data: Dataset, species: str, to_scale: str) -> Dataset:
 
 
 def _create_keyword_string(**kwargs: Any) -> str:
-    """
-    Create a formatted string for supplied keyword values. This will ignore
+    """Create a formatted string for supplied keyword values. This will ignore
     keywords where the value is None.
     This is used for printing details of keywords passed to the search functions.
     """
@@ -734,7 +784,7 @@ def _create_keyword_string(**kwargs: Any) -> str:
 
 def _metadata_difference(
     data: multDataTypes, params: Optional[list] = None, print_output: bool = True
-) -> Dict[str, list]:
+) -> dict[str, list]:
     """Check differences between metadata for returned data objects. Note this will
     only look at differences between values which are strings (not lists, floats etc.)
 
@@ -742,6 +792,7 @@ def _metadata_difference(
         data: Multiple data objects e.g. multiple ObsData as a list
         params: Specific metadata parameters to check. If None all parameters will be checked
         print_output: Summarise and print output to screen.
+
     Returns:
         Dict[str, list]: Keys and lists of values from the metadata with differences.
     """
@@ -789,16 +840,13 @@ def _metadata_difference(
     #         continue
 
     # - Collate summary of differences as a dictionary which maps as param: list of values
-    summary_difference: Dict[str, list] = {}
+    summary_difference: dict[str, list] = {}
     for param in param_difference:
         summary_difference[param] = []
         if print_output:
             logger.info(f" {param}: ")
         for m in metadata:
-            if param in m:
-                value = m[param]
-            else:
-                value = "NOT PRESENT"
+            value = m.get(param, "NOT PRESENT")
             summary_difference[param].append(value)
             if print_output:
                 logger.info(f" '{value}', ")
@@ -819,8 +867,7 @@ def _metadata_difference(
 def _metadata_difference_formatted(
     data: multDataTypes, params: Optional[list] = None, print_output: bool = True
 ) -> str:
-    """
-    Create formatted string for the difference in metadata between input objects.
+    """Create formatted string for the difference in metadata between input objects.
 
     Args:
         data : Multiple data objects e.g. multiple ObsData as a list