+++ openghg/util/_inlet.py
@@ -26,12 +26,30 @@ def format_inlet(
 ) -> None: ...
 
 
+@overload
 def format_inlet(
-    inlet: Optional[str],
+    inlet: slice,
     units: str = "m",
     key_name: Optional[str] = None,
     special_keywords: Optional[list] = None,
-) -> Optional[str]:
+) -> slice: ...
+
+
+@overload
+def format_inlet(
+    inlet: list[Union[str, slice, None]],
+    units: str = "m",
+    key_name: Optional[str] = None,
+    special_keywords: Optional[list] = None,
+) -> list[Union[str, slice, None]]: ...
+
+
+def format_inlet(
+    inlet: Union[str, slice, None, list[Union[str, slice, None]]],
+    units: str = "m",
+    key_name: Optional[str] = None,
+    special_keywords: Optional[list] = None,
+) -> Union[str, slice, None, list[Union[str, slice, None]]]:
     """
     Make sure inlet / height name conforms to standard. The standard
     imposed can depend on the associated key_name itself (can
@@ -55,7 +73,7 @@ def format_inlet(
             If so do not apply any formatting.
             If this is not set a special keyword of "multiple" and "column" will still be allowed.
     Returns:
-        str: formatted inlet string / None
+        same type as input, with all strings formatted
 
     Usage:
         >>> format_inlet("10")
@@ -66,6 +84,8 @@ def format_inlet(
             "10m"
         >>> format_inlet("10.111")
             "10.1m"
+        >>> format_inlet(["10", 100])
+            ["10m", "100m"]
         >>> format_inlet("multiple")
             "multiple"
         >>> format_inlet("10m", key_name="inlet")
@@ -75,8 +95,13 @@ def format_inlet(
         >>> format_inlet("10m", key_name="station_height_masl")
             "10"
     """
-    if inlet is None:
-        return None
+    # process list recursively
+    if isinstance(inlet, list):
+        return [format_inlet(x) for x in inlet]
+
+    # pass through None and slice
+    if inlet is None or isinstance(inlet, slice):
+        return inlet
 
     # By default the special keyword is "multiple" for data containing multiple inlets.
     # This will be included if data is a combined object from the object store.